bug fix for 64 bits platform
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@2002 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/include/tulip/AcyclicTest.h
@@ -59,7 +59,7 @@ private:
   void destroy(Graph *);
   AcyclicTest();
   static AcyclicTest * instance;
-  TLP_HASH_MAP<unsigned int,bool> resultsBuffer;
+  TLP_HASH_MAP<unsigned long, bool> resultsBuffer;
 };
 }
 /*@}*/
